Split this mega component up into multiple fiiles
